+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+04.13.2023
+- Added `--nophyloacc` option that prevents execution of the PhyloAcc rules in the snakemake workflow, useful for debugging or just running `--theta`
+- Internally, switched the number of informative sites required for a locus to be used in `--theta` estimation to be a param, maybe user option later
+
 04.06.2023
 - Fixed ids dir when `--theta` is set
 - Added `-iqtree-path` and `-coal-cmd` options to specify programs to estimate branch lengths in coalescent units when `--theta` is set